pi/ambient-temp-monitor/ambient-temp-logger/requirements.txt
influxdb-client
pyserial